+++ app/src/main/java/algonquin/cst2335/androidfinalproject/recipe/Constants.java
@@ -1,15 +1,5 @@
-/**
- * Filename: Constants.java
- * Purpose: Defines a class containing constant values used in the Recipe package.
- * Author: Liying Guo
- * Lab Section: CST2355 011
- * Creation Date: March 31, 2024
- */
 package algonquin.cst2335.androidfinalproject.recipe;
 
-/**
- * Class containing constant values API Key used in the Recipe package.
- */
 public class Constants {
-    public static final String RECEIPE_API="e848d564ec064a25b135e07444c5c790";//"c918ec43605843bfbbb1e58441d40b7c" "e848d564ec064a25b135e07444c5c790" "e162d89362c743cdbe534307d72f0335"
+    public static final String RECEIPE_API="40ef96e3f7a8460e822f68aca63d4995";//"c918ec43605843bfbbb1e58441d40b7c"
 }
