+++ app/src/main/java/algonquin/cst2335/androidfinalproject/song/DeezerApi.java
@@ -1,43 +1,70 @@
 package algonquin.cst2335.androidfinalproject.song;
+/*
+ * Filename: DeezerApi.java
+ * Author: Zhaoguo Han
+ * Lab Section: CST2355 011
+ * Creation Date: March 31, 2024
+ * Purpose: Handles communication with the Deezer API to search for artists and retrieve their songs.
+ */
 
 import android.content.Context;
-import android.util.Log;
-
 import com.android.volley.Request;
-import com.android.volley.RequestQueue;
 import com.android.volley.Response;
 import com.android.volley.VolleyError;
 import com.android.volley.toolbox.JsonObjectRequest;
-import com.android.volley.toolbox.Volley;
 
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.io.UnsupportedEncodingException;
+import java.net.URLEncoder;
 import java.util.ArrayList;
 import java.util.List;
-import java.net.URLEncoder;
-import android.util.Log;
 
+/**
+ * DeezerApi class handles the communication with the Deezer API to search for artists and retrieve their songs.
+ */
 public class DeezerApi {
 
     private static final String TAG = "DeezerApi";
     private static final String SEARCH_ARTIST_URL = "https://api.deezer.com/search/artist/?q=";
 
-    // Define the callback interface
+    /**
+     * Callback interface for handling asynchronous responses.
+     *
+     * @param <T> Type of the response object
+     */
     public interface Callback<T> {
+        /**
+         * Called when the request is successful.
+         *
+         * @param response The response object
+         */
         void onSuccess(T response);
+
+        /**
+         * Called when the request encounters an error.
+         *
+         * @param errorMessage The error message
+         */
         void onError(String errorMessage);
     }
 
+    /**
+     * Searches for an artist on Deezer API and retrieves their songs.
+     *
+     * @param context    The context of the calling activity
+     * @param artistName The name of the artist to search for
+     * @param callback   The callback to handle the response
+     */
     public static void searchArtist(Context context, String artistName, final Callback<List<Song>> callback) {
         try {
             // Encode artistName
             String encodedArtistName = URLEncoder.encode(artistName, "UTF-8");
 
             // Construct the URL
-            String url = "https://api.deezer.com/search/artist/?q=" + encodedArtistName;
+            String url = SEARCH_ARTIST_URL + encodedArtistName;
 
             // Create a JsonObjectRequest to make the network request
             JsonObjectRequest jsonObjectRequest = new JsonObjectRequest(Request.Method.GET, url, null,
@@ -61,7 +88,7 @@ public class DeezerApi {
                                                     JSONArray tracks = null;
                                                     try {
                                                         tracks = response1.getJSONArray("data");
-                                                        for (int k = 0; k < tracks.length(); k++){
+                                                        for (int k = 0; k < tracks.length(); k++) {
                                                             JSONObject song = tracks.getJSONObject(k);
                                                             String title = song.getString("title");
                                                             int duration = song.getInt("duration");
@@ -109,38 +136,4 @@ public class DeezerApi {
             throw new RuntimeException(e);
         }
     }
-
-
-    // Method to parse the response and extract the list of songs
-
-
-    private static List<Song> parseSongsFromResponse(JSONObject response) {
-        List<Song> songs = new ArrayList<>();
-
-        try {
-            JSONArray data = response.getJSONArray("data");
-
-            for (int i = 0; i < data.length(); i++) {
-                JSONObject songObject = data.getJSONObject(i);
-
-                // Extract song details from JSON object
-                String title = songObject.getString("title");
-                String album = songObject.getJSONObject("album").getString("title");
-                       album = songObject.getString("name");
-
-                int duration = songObject.getInt("duration");
-                String cover = songObject.getString("cover");
-
-                // Create Song object and add it to the list
-                Song song = new Song(title, album, duration, cover);
-                songs.add(song);
-            }
-        } catch (JSONException e) {
-            e.printStackTrace();
-        }
-
-        return songs;
-    }
-
-
 }
