+++ app/src/main/java/algonquin/cst2335/androidfinalproject/song/SongAdapter.java
@@ -1,3 +1,10 @@
+/*
+ * Filename: SongAdapter.java
+ * Author: Zhaoguo Han
+ * Lab Section: CST2355 011
+ * Creation Date: March 31, 2024
+ * Purpose: This class is an adapter for the RecyclerView in the MainSongActivity. It is responsible for displaying song data and handling user interactions with the song items.
+ */
 package algonquin.cst2335.androidfinalproject.song;
 
 import android.view.LayoutInflater;
@@ -16,24 +23,37 @@ import java.util.List;
 
 import algonquin.cst2335.androidfinalproject.R;
 
-
+/**
+ * Adapter class for RecyclerView to display songs.
+ */
 public class SongAdapter extends RecyclerView.Adapter<SongAdapter.SongViewHolder> {
 
     private List<Song> songList;
     private SongDao songDao;
     private boolean showDeleteButton = false;
     private boolean showAddButton = true;
-    private OnFavoriteClickListener onFavoriteClickListener; // nterface variable
+    private OnFavoriteClickListener onFavoriteClickListener;
 
+    /**
+     * Setter method to set the list of songs.
+     * @param songList The list of songs to be displayed.
+     */
     public void setSongList(List<Song> songList) {
         this.songList = songList;
         notifyDataSetChanged();
     }
-    public void setSongDao(SongDao songDao)
-    {
+
+    /**
+     * Setter method to set the DAO for accessing the database.
+     * @param songDao The SongDao object.
+     */
+    public void setSongDao(SongDao songDao) {
         this.songDao = songDao;
     }
 
+    /**
+     * Create a SongViewHolder instance.
+     */
     @NonNull
     @Override
     public SongViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
@@ -41,6 +61,9 @@ public class SongAdapter extends RecyclerView.Adapter<SongAdapter.SongViewHolder
         return new SongViewHolder(view);
     }
 
+    /**
+     * Bind data to the ViewHolder.
+     */
     @Override
     public void onBindViewHolder(@NonNull SongViewHolder holder, int position) {
         Song song = songList.get(position);
@@ -52,40 +75,34 @@ public class SongAdapter extends RecyclerView.Adapter<SongAdapter.SongViewHolder
         holder.buttonAddToFavorites.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                // 获取当前位置的歌曲并插入到数据库中
                 int adapterPosition = holder.getAdapterPosition();
                 if (adapterPosition != RecyclerView.NO_POSITION) {
                     Song selectedSong = songList.get(adapterPosition);
                     insertSongToDatabase(selectedSong);
                 }
-                if (onFavoriteClickListener!= null) {
+                if (onFavoriteClickListener != null) {
                     onFavoriteClickListener.onFavoriteClick(song);
                 }
             }
         });
 
-        // Set click listener for the delete button
         holder.buttonDeleteFavorite.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                // Get the position of the song in the list
                 int adapterPosition = holder.getAdapterPosition();
                 if (adapterPosition != RecyclerView.NO_POSITION) {
-                    // Get the selected song
                     Song selectedSong = songList.get(adapterPosition);
-                    // Call the deleteFavoriteSong method in the activity
                     ((MainSongActivity) v.getContext()).deleteFavoriteSong(selectedSong);
                 }
             }
         });
 
-        // 设置删除按钮的可见性
         if (showDeleteButton) {
             holder.buttonDeleteFavorite.setVisibility(View.VISIBLE);
         } else {
             holder.buttonDeleteFavorite.setVisibility(View.GONE);
         }
-        // 设置删除按钮的可见性
+
         if (showAddButton) {
             holder.buttonAddToFavorites.setVisibility(View.VISIBLE);
         } else {
@@ -93,16 +110,21 @@ public class SongAdapter extends RecyclerView.Adapter<SongAdapter.SongViewHolder
         }
     }
 
+    /**
+     * Get the number of items in the RecyclerView.
+     */
     @Override
     public int getItemCount() {
         return songList != null ? songList.size() : 0;
     }
 
+    /**
+     * ViewHolder class to hold the views for each item in the RecyclerView.
+     */
     static class SongViewHolder extends RecyclerView.ViewHolder {
         TextView textViewTitle, textViewAlbum, textViewDuration;
         ImageView imageViewAlbumCover;
-        Button buttonAddToFavorites;
-        Button buttonDeleteFavorite;
+        Button buttonAddToFavorites, buttonDeleteFavorite;
 
         public SongViewHolder(@NonNull View itemView) {
             super(itemView);
@@ -110,20 +132,27 @@ public class SongAdapter extends RecyclerView.Adapter<SongAdapter.SongViewHolder
             textViewAlbum = itemView.findViewById(R.id.textViewAlbum);
             textViewDuration = itemView.findViewById(R.id.textViewDuration);
             imageViewAlbumCover = itemView.findViewById(R.id.imageViewAlbumCover);
-            buttonAddToFavorites = itemView.findViewById(R.id.buttonAddToFavorites); // Add this line to initialize buttonAddToFavorites
+            buttonAddToFavorites = itemView.findViewById(R.id.buttonAddToFavorites);
             buttonDeleteFavorite = itemView.findViewById(R.id.buttonDeleteFavorite);
         }
     }
+
+    /**
+     * Interface definition for a callback to be invoked when a song is added to favorites.
+     */
     public interface OnFavoriteClickListener {
         void onFavoriteClick(Song song);
     }
-    // Method to set the listener
+
+    /**
+     * Method to set the listener for favorite button clicks.
+     * @param listener The listener to be set.
+     */
     public void setOnFavoriteClickListener(OnFavoriteClickListener listener) {
-        this.onFavoriteClickListener= listener;
+        this.onFavoriteClickListener = listener;
     }
 
     private void insertSongToDatabase(Song song) {
-        // spcifying data insert to database
         if (songDao != null) {
             new Thread(new Runnable() {
                 @Override
@@ -134,16 +163,28 @@ public class SongAdapter extends RecyclerView.Adapter<SongAdapter.SongViewHolder
         }
     }
 
-    // Add a method to set the list of favorite songs
+    /**
+     * Method to set the list of favorite songs.
+     * @param favoriteSongs The list of favorite songs.
+     */
     public void setFavoriteSongs(List<Song> favoriteSongs) {
         this.songList = favoriteSongs;
         notifyDataSetChanged();
     }
 
+    /**
+     * Method to show or hide the delete button.
+     * @param show true to show the delete button, false to hide.
+     */
     public void setShowDeleteButton(boolean show) {
         this.showDeleteButton = show;
         notifyDataSetChanged();
     }
+
+    /**
+     * Method to show or hide the add button.
+     * @param show true to show the add button, false to hide.
+     */
     public void setShowAddButton(boolean show) {
         this.showAddButton = show;
         notifyDataSetChanged();
